Finance frontend
change side bar and add finance pages
