+++ frontend/package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^7.0.0",
-    "@tanstack/react-query": "^5.85.0",
+    "@tanstack/react-query": "^5.85.3",
+    "@tanstack/react-query-devtools": "^5.85.3",
     "axios": "^1.11.0",
     "framer-motion": "^12.23.9",
     "lucide-react": "^0.539.0",
