+++ frontend/src/App.jsx
@@ -9,13 +9,15 @@ import AboutUs from "./pages/aboutus";
 import ContactUs from "./pages/contactus";
 import UserProfile from "./pages/UserProfile";
 
-import FinanceDashboard from "./admin/FinanceDashboard";
+import FinanceOverview from "./admin/FinanceOverview";
+import FinanceTransaction from "./admin/FinanceTransaction";
+import FinanceNewTransaction from "./admin/FinanceNewTransaction";
+
 import CheckoutPage from "./pages/checkout"; // Correct import name
 // FIX: The path you had was './pages/store/...', let's assume it's directly in './pages/'
-import OrderSuccessPage from './pages/store/OrderSuccessPage';
-import OrderCancelPage from './pages/store/OrderCancelPage';
-import MyOrdersPage from './pages/MyOrdersPage'; 
-
+import OrderSuccessPage from "./pages/store/OrderSuccessPage";
+import OrderCancelPage from "./pages/store/OrderCancelPage";
+import MyOrdersPage from "./pages/MyOrdersPage";
 
 // --- ADMIN PAGES (Lazy Loaded & Direct Imports) ---
 const AdminLayout = lazy(() => import("./admin/AdminLayout"));
@@ -25,20 +27,32 @@ const StoreDashboard = lazy(() => import("./admin/AdminDashboard"));
 import AdminProducts from "./pages/store/Products";
 import AdminOrders from "./pages/store/Orders";
 
-
-
 // --- PLACEHOLDER COMPONENTS ---
 // These are fine as they are. They are simple, non-lazy components.
-const FarmDashboard = () => <div className="p-6 text-2xl font-bold">Farm Overview Dashboard</div>;
-const LivestockPage = () => <div className="p-6 text-2xl font-bold">Livestock Management</div>;
-const CropPage = () => <div className="p-6 text-2xl font-bold">Crop Management</div>;
-const StaffPage = () => <div className="p-6 text-2xl font-bold">Staff Management</div>;
-const RevenuePage = () => <div className="p-6 text-2xl font-bold">Revenue & Financials</div>;
-const DiscountsPage = () => <div className="p-6 text-2xl font-bold">Discount Management</div>;
-const CustomersPage = () => <div className="p-6 text-2xl font-bold">Customer Management</div>;
-const ReportsPage = () => <div className="p-6 text-2xl font-bold">Store Reports</div>;
-
-
+const FarmDashboard = () => (
+  <div className="p-6 text-2xl font-bold">Farm Overview Dashboard</div>
+);
+const LivestockPage = () => (
+  <div className="p-6 text-2xl font-bold">Livestock Management</div>
+);
+const CropPage = () => (
+  <div className="p-6 text-2xl font-bold">Crop Management</div>
+);
+const StaffPage = () => (
+  <div className="p-6 text-2xl font-bold">Staff Management</div>
+);
+const RevenuePage = () => (
+  <div className="p-6 text-2xl font-bold">Revenue & Financials</div>
+);
+const DiscountsPage = () => (
+  <div className="p-6 text-2xl font-bold">Discount Management</div>
+);
+const CustomersPage = () => (
+  <div className="p-6 text-2xl font-bold">Customer Management</div>
+);
+const ReportsPage = () => (
+  <div className="p-6 text-2xl font-bold">Store Reports</div>
+);
 
 // --- Main App Component ---
 export default function App() {
@@ -51,15 +65,14 @@ export default function App() {
         <Route path="/about" element={<AboutUs />} />
         <Route path="/contact" element={<ContactUs />} />
         <Route path="/profile" element={<UserProfile />} />
-        <Route path="/checkout" element={<CheckoutPage />} /> 
+        <Route path="/checkout" element={<CheckoutPage />} />
         <Route path="/my-orders" element={<MyOrdersPage />} />
-        
+
         {/* The order completion pages also use the main layout */}
         <Route path="/order/success" element={<OrderSuccessPage />} />
         <Route path="/order/cancel" element={<OrderCancelPage />} />
       </Route>
 
-      
       {/* === ADMIN ROUTES (Wrapped with special AdminLayout) === */}
       <Route
         path="/admin"
@@ -82,23 +95,33 @@ export default function App() {
         <Route path="staff" element={<StaffPage />} />
         <Route path="revenue" element={<RevenuePage />} />
 
-
         {/* --- Nested Store Management Routes --- */}
-        <Route path="store/dashboard" element={ <Suspense fallback={<div className="p-6">Loading Dashboard…</div>}><StoreDashboard /></Suspense> } />
+        <Route
+          path="store/dashboard"
+          element={
+            <Suspense fallback={<div className="p-6">Loading Dashboard…</div>}>
+              <StoreDashboard />
+            </Suspense>
+          }
+        />
         <Route path="store/products" element={<AdminProducts />} />
         <Route path="store/orders" element={<AdminOrders />} />
 
         <Route path="store/discounts" element={<DiscountsPage />} />
         <Route path="store/customers" element={<CustomersPage />} />
         <Route path="store/reports" element={<ReportsPage />} />
+
+        {/* Finance */}
+        <Route path="finance">
+          <Route index element={<Navigate to="overview" replace />} />
+          <Route path="overview" element={<FinanceOverview />} />
+          <Route path="transaction" element={<FinanceTransaction />} />
+          <Route path="new_transaction" element={<FinanceNewTransaction />} />
+        </Route>
       </Route>
-
-
-      <Route path="/finance" element={<FinanceDashboard />} />
 
       {/* === 404 FALLBACK ROUTE === */}
 
-      
       {/* === FALLBACK ROUTE (Catches any unmatched URL) === */}
 
       <Route path="*" element={<Navigate to="/" replace />} />
