--- frontend/src/admin/FinanceDashboard.jsx
@@ -1,1255 +1,0 @@
-import React, { useState, useMemo } from "react";
-import {
-  BarChart,
-  Bar,
-  XAxis,
-  YAxis,
-  Tooltip,
-  ResponsiveContainer,
-  CartesianGrid,
-  LineChart,
-  Line,
-  PieChart,
-  Pie,
-  Cell,
-} from "recharts";
-
-const theme = {
-  primary: {
-    300: "#86efac",
-    400: "#4ade80",
-    500: "#22c55e",
-    600: "#16a34a",
-  },
-  secondary: {
-    500: "#ef4444",
-  },
-  neutral: {
-    200: "#e5e7eb",
-    300: "#d1d5db",
-    400: "#9ca3af",
-    600: "#4b5563",
-  },
-};
-
-function currency(n) {
-  if (isNaN(n)) return "—";
-  return n.toLocaleString("en-LK", {
-    style: "currency",
-    currency: "LKR",
-    maximumFractionDigits: 0,
-  });
-}
-
-function monthKey(iso) {
-  return (iso || "").slice(0, 7);
-}
-
-function downloadCSV(filename, rows) {
-  const headers = ["id", "type", "date", "category", "amount", "note"];
-  const body = rows.map((r) =>
-    headers
-      .map((h) => String(r[h] ?? "").replace(/"/g, '""'))
-      .map((v) => `"${v}"`)
-      .join(",")
-  );
-  const csv = [headers.join(","), ...body].join("\n");
-  const blob = new Blob([csv], { type: "text/csv;charset=utf-8;" });
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement("a");
-  a.href = url;
-  a.download = filename;
-  a.click();
-  URL.revokeObjectURL(url);
-}
-
-const StatCard = ({ title, value, change, icon, trend }) => (
-  <div className="bg-white rounded-xl p-6 border border-gray-100 shadow-lg hover:shadow-xl transition-shadow duration-300">
-    <div className="flex items-center justify-between">
-      <div className="space-y-2">
-        <p className="text-sm font-medium text-gray-600">{title}</p>
-        <p className="text-2xl font-bold text-gray-900">{value}</p>
-        {change && (
-          <div
-            className={`flex items-center text-sm ${
-              trend === "up"
-                ? "text-emerald-600"
-                : trend === "down"
-                ? "text-red-600"
-                : "text-gray-500"
-            }`}
-          >
-            {trend === "up" && (
-              <svg
-                className="w-4 h-4 mr-1"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M7 17l9.2-9.2M17 17V7M17 17H7"
-                />
-              </svg>
-            )}
-            {trend === "down" && (
-              <svg
-                className="w-4 h-4 mr-1"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M7 7l9.2 9.2M17 7v10M17 7H7"
-                />
-              </svg>
-            )}
-            {change}
-          </div>
-        )}
-      </div>
-      <div className="h-12 w-12 rounded-lg bg-gradient-to-br from-emerald-50 to-emerald-100 flex items-center justify-center">
-        {icon}
-      </div>
-    </div>
-  </div>
-);
-
-// Overview Page Component
-const OverviewPage = ({ transactions = [] }) => {
-  const stats = useMemo(() => {
-    const totalIncome = transactions
-      .filter((r) => r.type === "income")
-      .reduce((sum, r) => sum + r.amount, 0);
-    const totalExpenses = transactions
-      .filter((r) => r.type === "expense")
-      .reduce((sum, r) => sum + r.amount, 0);
-    const netProfit = totalIncome - totalExpenses;
-
-    return {
-      totalIncome,
-      totalExpenses,
-      netProfit,
-      transactionCount: transactions.length,
-    };
-  }, [transactions]);
-
-  const chartData = useMemo(() => {
-    const map = new Map();
-    transactions.forEach((r) => {
-      const key = monthKey(r.date);
-      if (!key) return;
-      if (!map.has(key)) map.set(key, { month: key, income: 0, expense: 0 });
-      map.get(key)[r.type === "income" ? "income" : "expense"] += r.amount;
-    });
-    return Array.from(map.values()).sort((a, b) =>
-      a.month.localeCompare(b.month)
-    );
-  }, [transactions]);
-
-  const cumulativeData = useMemo(() => {
-    let running = 0;
-    return chartData.map((d) => {
-      running += d.income - d.expense;
-      return { month: d.month, net: running };
-    });
-  }, [chartData]);
-
-  const categoryData = useMemo(() => {
-    const map = new Map();
-    transactions.forEach((r) => {
-      if (!map.has(r.category)) {
-        map.set(r.category, { name: r.category, value: 0, type: r.type });
-      }
-      map.get(r.category).value += r.amount;
-    });
-    return Array.from(map.values())
-      .sort((a, b) => b.value - a.value)
-      .slice(0, 6);
-  }, [transactions]);
-
-  const pieColors = [
-    theme.primary[500],
-    theme.primary[400],
-    theme.primary[300],
-    theme.secondary[500],
-    theme.neutral[400],
-    theme.neutral[300],
-  ];
-
-  return (
-    <div className="space-y-8">
-      {/* Stats Grid */}
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
-        <StatCard
-          title="Total Income"
-          value={currency(stats.totalIncome)}
-          change="+12.5% from last month"
-          trend="up"
-          icon={
-            <svg
-              className="w-6 h-6 text-emerald-600"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M7 17l9.2-9.2M17 17V7M17 17H7"
-              />
-            </svg>
-          }
-        />
-        <StatCard
-          title="Total Expenses"
-          value={currency(stats.totalExpenses)}
-          change="+8.2% from last month"
-          trend="up"
-          icon={
-            <svg
-              className="w-6 h-6 text-red-600"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M7 7l9.2 9.2M17 7v10M17 7H7"
-              />
-            </svg>
-          }
-        />
-        <StatCard
-          title="Net Profit"
-          value={currency(stats.netProfit)}
-          change={
-            stats.netProfit >= 0 ? "+15.3% profit margin" : "Loss this period"
-          }
-          trend={stats.netProfit >= 0 ? "up" : "down"}
-          icon={
-            <svg
-              className="w-6 h-6 text-emerald-600"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"
-              />
-            </svg>
-          }
-        />
-        <StatCard
-          title="Transactions"
-          value={stats.transactionCount.toString()}
-          change="Latest: Today"
-          icon={
-            <svg
-              className="w-6 h-6 text-emerald-600"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M9 5H7a2 2 0 00-2 2v10a2 2 0 002 2h8a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2"
-              />
-            </svg>
-          }
-        />
-      </div>
-
-      {/* Charts Grid */}
-      <div className="grid lg:grid-cols-3 gap-8">
-        <div className="lg:col-span-2 bg-white rounded-xl p-6 border border-gray-100 shadow-lg hover:shadow-xl transition-shadow duration-300">
-          <h3 className="text-lg font-semibold text-gray-900 mb-4">
-            Monthly Income vs Expense
-          </h3>
-          <div className="h-80">
-            <ResponsiveContainer width="100%" height="100%">
-              <BarChart
-                data={chartData}
-                margin={{ top: 20, right: 30, left: 20, bottom: 5 }}
-              >
-                <CartesianGrid
-                  strokeDasharray="3 3"
-                  stroke={theme.neutral[200]}
-                />
-                <XAxis
-                  dataKey="month"
-                  stroke={theme.neutral[600]}
-                  fontSize={12}
-                />
-                <YAxis stroke={theme.neutral[600]} fontSize={12} />
-                <Tooltip
-                  formatter={(v) => [currency(Number(v)), ""]}
-                  labelFormatter={(label) => `Month: ${label}`}
-                  contentStyle={{
-                    backgroundColor: "white",
-                    border: `1px solid ${theme.neutral[200]}`,
-                    borderRadius: "8px",
-                    boxShadow: "0 4px 6px -1px rgb(0 0 0 / 0.1)",
-                  }}
-                />
-                <Bar
-                  dataKey="income"
-                  name="Income"
-                  fill={theme.primary[500]}
-                  radius={[4, 4, 0, 0]}
-                />
-                <Bar
-                  dataKey="expense"
-                  name="Expense"
-                  fill={theme.secondary[500]}
-                  radius={[4, 4, 0, 0]}
-                />
-              </BarChart>
-            </ResponsiveContainer>
-          </div>
-        </div>
-
-        <div className="bg-white rounded-xl p-6 border border-gray-100 shadow-lg hover:shadow-xl transition-shadow duration-300">
-          <h3 className="text-lg font-semibold text-gray-900 mb-4">
-            Top Categories
-          </h3>
-          <div className="h-80">
-            <ResponsiveContainer width="100%" height="100%">
-              <PieChart>
-                <Pie
-                  data={categoryData}
-                  cx="50%"
-                  cy="50%"
-                  innerRadius={60}
-                  outerRadius={100}
-                  paddingAngle={5}
-                  dataKey="value"
-                >
-                  {categoryData.map((entry, index) => (
-                    <Cell
-                      key={`cell-${index}`}
-                      fill={pieColors[index % pieColors.length]}
-                    />
-                  ))}
-                </Pie>
-                <Tooltip formatter={(v) => currency(Number(v))} />
-              </PieChart>
-            </ResponsiveContainer>
-          </div>
-          <div className="mt-4 space-y-2">
-            {categoryData.slice(0, 4).map((item, index) => (
-              <div
-                key={item.name}
-                className="flex items-center justify-between text-sm"
-              >
-                <div className="flex items-center space-x-2">
-                  <div
-                    className="w-3 h-3 rounded-full"
-                    style={{ backgroundColor: pieColors[index] }}
-                  />
-                  <span className="text-gray-600">{item.name}</span>
-                </div>
-                <span className="font-medium text-gray-900">
-                  {currency(item.value)}
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
-
-      {/* Cumulative Net Chart */}
-      <div className="bg-white rounded-xl p-6 border border-gray-100 shadow-lg hover:shadow-xl transition-shadow duration-300">
-        <h3 className="text-lg font-semibold text-gray-900 mb-4">
-          Cumulative Net Balance
-        </h3>
-        <div className="h-80">
-          <ResponsiveContainer width="100%" height="100%">
-            <LineChart
-              data={cumulativeData}
-              margin={{ top: 20, right: 30, left: 20, bottom: 5 }}
-            >
-              <CartesianGrid
-                strokeDasharray="3 3"
-                stroke={theme.neutral[200]}
-              />
-              <XAxis
-                dataKey="month"
-                stroke={theme.neutral[600]}
-                fontSize={12}
-              />
-              <YAxis stroke={theme.neutral[600]} fontSize={12} />
-              <Tooltip
-                formatter={(v) => [currency(Number(v)), "Net Balance"]}
-                labelFormatter={(label) => `Month: ${label}`}
-                contentStyle={{
-                  backgroundColor: "white",
-                  border: `1px solid ${theme.neutral[200]}`,
-                  borderRadius: "8px",
-                  boxShadow: "0 4px 6px -1px rgb(0 0 0 / 0.1)",
-                }}
-              />
-              <Line
-                type="monotone"
-                dataKey="net"
-                stroke={theme.primary[500]}
-                strokeWidth={3}
-                dot={{ fill: theme.primary[500], strokeWidth: 2, r: 4 }}
-                activeDot={{ r: 6, fill: theme.primary[600] }}
-              />
-            </LineChart>
-          </ResponsiveContainer>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-// Transactions Page Component
-const TransactionsPage = ({
-  transactions = [],
-  onEdit,
-  onDelete,
-  onAddNew,
-}) => {
-  const [q, setQ] = useState("");
-  const [typeFilter, setTypeFilter] = useState("all");
-  const [monthFilter, setMonthFilter] = useState("all");
-  const [showDeleteModal, setShowDeleteModal] = useState(null);
-
-  const months = useMemo(() => {
-    const set = new Set(transactions.map((r) => monthKey(r.date)));
-    return ["all", ...Array.from(set).sort().reverse()];
-  }, [transactions]);
-
-  const filtered = transactions.filter((r) => {
-    if (typeFilter !== "all" && r.type !== typeFilter) return false;
-    if (monthFilter !== "all" && monthKey(r.date) !== monthFilter) return false;
-    if (!q) return true;
-    const hay = `${r.category} ${r.note} ${r.type} ${r.date}`.toLowerCase();
-    return hay.includes(q.toLowerCase());
-  });
-
-  const handleDelete = (id) => {
-    onDelete(id);
-    setShowDeleteModal(null);
-  };
-
-  return (
-    <div className="space-y-6">
-      {/* Filters */}
-      <div className="bg-gradient-to-r from-white to-emerald-50 rounded-2xl shadow-xl border border-emerald-100 p-8 hover:shadow-2xl transition-all duration-300 transform hover:-translate-y-1">
-        <div className="flex items-center mb-6">
-          <div className="h-10 w-10 rounded-xl bg-gradient-to-br from-emerald-500 to-emerald-600 flex items-center justify-center mr-3 shadow-lg">
-            <svg
-              className="w-5 h-5 text-white"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M3 4a1 1 0 011-1h16a1 1 0 011 1v2.586a1 1 0 01-.293.707l-6.414 6.414a1 1 0 00-.293.707V17l-4 4v-6.586a1 1 0 00-.293-.707L3.293 7.293A1 1 0 013 6.586V4z"
-              />
-            </svg>
-          </div>
-          <h3 className="text-xl font-bold text-gray-900">Filter & Search</h3>
-        </div>
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-          <div className="space-y-2">
-            <label className="block text-sm font-semibold text-gray-700">
-              Search Transactions
-            </label>
-            <div className="relative">
-              <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                <svg
-                  className="h-5 w-5 text-gray-400"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                  />
-                </svg>
-              </div>
-              <input
-                placeholder="Search category, note, date…"
-                className="w-full pl-10 pr-4 py-3 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-inner bg-white/80 backdrop-blur-sm transition-all duration-200"
-                value={q}
-                onChange={(e) => setQ(e.target.value)}
-              />
-            </div>
-          </div>
-          <div className="space-y-2">
-            <label className="block text-sm font-semibold text-gray-700">
-              Transaction Type
-            </label>
-            <div className="relative">
-              <select
-                className="w-full py-3 px-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-inner bg-white/80 backdrop-blur-sm transition-all duration-200 appearance-none"
-                value={typeFilter}
-                onChange={(e) => setTypeFilter(e.target.value)}
-              >
-                <option value="all">🔄 All Types</option>
-                <option value="income">💰 Income Only</option>
-                <option value="expense">💸 Expenses Only</option>
-              </select>
-              <div className="absolute inset-y-0 right-0 flex items-center pr-3 pointer-events-none">
-                <svg
-                  className="h-5 w-5 text-gray-400"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M19 9l-7 7-7-7"
-                  />
-                </svg>
-              </div>
-            </div>
-          </div>
-          <div className="space-y-2">
-            <label className="block text-sm font-semibold text-gray-700">
-              Filter by Month
-            </label>
-            <div className="relative">
-              <select
-                className="w-full py-3 px-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-inner bg-white/80 backdrop-blur-sm transition-all duration-200 appearance-none"
-                value={monthFilter}
-                onChange={(e) => setMonthFilter(e.target.value)}
-              >
-                {months.map((m) => (
-                  <option key={m} value={m}>
-                    {m === "all" ? "📅 All Months" : `📅 ${m}`}
-                  </option>
-                ))}
-              </select>
-              <div className="absolute inset-y-0 right-0 flex items-center pr-3 pointer-events-none">
-                <svg
-                  className="h-5 w-5 text-gray-400"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M19 9l-7 7-7-7"
-                  />
-                </svg>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Transactions Table */}
-      <div className="bg-gradient-to-r from-white to-gray-50 rounded-2xl shadow-2xl border border-gray-200 overflow-hidden hover:shadow-3xl transition-all duration-300 transform hover:-translate-y-1">
-        <div className="px-8 py-6 border-b border-gray-200 bg-gradient-to-r from-emerald-50 to-white flex items-center justify-between">
-          <div className="flex items-center space-x-4">
-            <div className="h-10 w-10 rounded-xl bg-gradient-to-br from-emerald-500 to-emerald-600 flex items-center justify-center shadow-lg">
-              <svg
-                className="w-5 h-5 text-white"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M9 5H7a2 2 0 00-2 2v10a2 2 0 002 2h8a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2"
-                />
-              </svg>
-            </div>
-            <div>
-              <h2 className="text-xl font-bold text-gray-900">
-                Transaction History
-              </h2>
-              <p className="text-sm text-gray-600 mt-1">
-                {filtered.length} records found
-              </p>
-            </div>
-          </div>
-          <div className="flex space-x-3">
-            <button
-              onClick={() => downloadCSV("finance-report.csv", filtered)}
-              className="inline-flex items-center px-4 py-2 border border-gray-300 rounded-lg text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-            >
-              <svg
-                className="w-4 h-4 mr-2"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M12 10v6m0 0l-3-3m3 3l3-3m2 8H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z"
-                />
-              </svg>
-              Export CSV
-            </button>
-            <button
-              onClick={onAddNew}
-              className="inline-flex items-center px-6 py-3 bg-gradient-to-r from-emerald-600 to-emerald-700 hover:from-emerald-700 hover:to-emerald-800 text-white rounded-xl text-sm font-semibold transition-all duration-200 shadow-lg hover:shadow-xl transform hover:-translate-y-0.5"
-            >
-              <svg
-                className="w-4 h-4 mr-2"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M12 4v16m8-8H4"
-                />
-              </svg>
-              Add New
-            </button>
-          </div>
-        </div>
-
-        <div className="overflow-x-auto">
-          <table className="min-w-full divide-y divide-gray-200">
-            <thead className="bg-gray-50">
-              <tr>
-                <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                  Date
-                </th>
-                <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                  Type
-                </th>
-                <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                  Category
-                </th>
-                <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                  Note
-                </th>
-                <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                  Amount
-                </th>
-                <th className="px-6 py-3 text-right text-xs font-medium text-gray-500 uppercase tracking-wider">
-                  Actions
-                </th>
-              </tr>
-            </thead>
-            <tbody className="bg-white divide-y divide-gray-200">
-              {filtered.length === 0 && (
-                <tr>
-                  <td colSpan={6} className="px-6 py-12 text-center">
-                    <div className="flex flex-col items-center">
-                      <svg
-                        className="w-12 h-12 text-gray-400 mb-4"
-                        fill="none"
-                        viewBox="0 0 24 24"
-                        stroke="currentColor"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M9 5H7a2 2 0 00-2 2v10a2 2 0 002 2h8a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2"
-                        />
-                      </svg>
-                      <p className="text-gray-500 text-lg">
-                        No transactions found
-                      </p>
-                      <p className="text-gray-400 text-sm">
-                        Try adjusting your filters or add a new transaction
-                      </p>
-                    </div>
-                  </td>
-                </tr>
-              )}
-              {filtered.map((r) => (
-                <tr key={r.id} className="hover:bg-gray-50 transition-colors">
-                  <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                    {r.date}
-                  </td>
-                  <td className="px-6 py-4 whitespace-nowrap">
-                    <span
-                      className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${
-                        r.type === "income"
-                          ? "bg-emerald-100 text-emerald-800"
-                          : "bg-red-100 text-red-800"
-                      }`}
-                    >
-                      {r.type === "income" ? "💰" : "💸"} {r.type}
-                    </span>
-                  </td>
-                  <td className="px-6 py-4 text-sm text-gray-900">
-                    {r.category}
-                  </td>
-                  <td className="px-6 py-4 text-sm text-gray-500 max-w-xs truncate">
-                    {r.note || "—"}
-                  </td>
-                  <td className="px-6 py-4 whitespace-nowrap text-sm font-medium text-gray-900">
-                    {currency(r.amount)}
-                  </td>
-                  <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
-                    <div className="flex justify-end space-x-2">
-                      <button
-                        onClick={() => onEdit(r)}
-                        className="inline-flex items-center px-3 py-1.5 border border-gray-300 rounded-md text-xs font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-                      >
-                        <svg
-                          className="w-3 h-3 mr-1"
-                          fill="none"
-                          viewBox="0 0 24 24"
-                          stroke="currentColor"
-                        >
-                          <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth={2}
-                            d="M11 5H6a2 2 0 00-2 2v11a2 2 0 002 2h11a2 2 0 002-2v-5m-1.414-9.414a2 2 0 112.828 2.828L11.828 15H9v-2.828l8.586-8.586z"
-                          />
-                        </svg>
-                        Edit
-                      </button>
-                      <button
-                        onClick={() => setShowDeleteModal(r.id)}
-                        className="inline-flex items-center px-3 py-1.5 border border-red-300 rounded-md text-xs font-medium text-red-700 bg-white hover:bg-red-50 transition-colors"
-                      >
-                        <svg
-                          className="w-3 h-3 mr-1"
-                          fill="none"
-                          viewBox="0 0 24 24"
-                          stroke="currentColor"
-                        >
-                          <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth={2}
-                            d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"
-                          />
-                        </svg>
-                        Delete
-                      </button>
-                    </div>
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      </div>
-
-      {/* Delete Confirmation Modal */}
-      {showDeleteModal && (
-        <div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full z-50 flex items-center justify-center">
-          <div className="bg-white p-6 rounded-xl shadow-2xl max-w-md w-full mx-4 transform transition-all duration-200">
-            <div className="flex items-center mb-4">
-              <div className="flex-shrink-0">
-                <svg
-                  className="w-6 h-6 text-red-600"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L4.268 16.5c-.77.833.192 2.5 1.732 2.5z"
-                  />
-                </svg>
-              </div>
-              <div className="ml-3">
-                <h3 className="text-lg font-medium text-gray-900">
-                  Delete Transaction
-                </h3>
-              </div>
-            </div>
-            <p className="text-sm text-gray-500 mb-6">
-              Are you sure you want to delete this transaction? This action
-              cannot be undone.
-            </p>
-            <div className="flex justify-end space-x-3">
-              <button
-                onClick={() => setShowDeleteModal(null)}
-                className="px-4 py-2 border border-gray-300 rounded-lg text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-              >
-                Cancel
-              </button>
-              <button
-                onClick={() => handleDelete(showDeleteModal)}
-                className="px-4 py-2 bg-red-600 hover:bg-red-700 text-white rounded-lg text-sm font-medium transition-colors"
-              >
-                Delete
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-// Add New Page Component
-const AddNewPage = ({ editingTransaction = null, onSave, onCancel }) => {
-  const [form, setForm] = useState({
-    type: "expense",
-    date: new Date().toISOString().slice(0, 10),
-    category: "",
-    amount: "",
-    note: "",
-  });
-
-  React.useEffect(() => {
-    if (editingTransaction) {
-      setForm({
-        type: editingTransaction.type,
-        date: editingTransaction.date,
-        category: editingTransaction.category,
-        amount: String(editingTransaction.amount),
-        note: editingTransaction.note ?? "",
-      });
-    } else {
-      setForm({
-        type: "expense",
-        date: new Date().toISOString().slice(0, 10),
-        category: "",
-        amount: "",
-        note: "",
-      });
-    }
-  }, [editingTransaction]);
-
-  function handleSubmit() {
-    const amt = Number(form.amount);
-    if (!form.date || !form.category || !amt || amt <= 0) {
-      alert("Please fill date, category and a positive amount.");
-      return;
-    }
-
-    const transaction = {
-      ...form,
-      amount: amt,
-      id: editingTransaction?.id || crypto.randomUUID(),
-    };
-
-    onSave(transaction, editingTransaction ? "edit" : "add");
-  }
-
-  function handleCancel() {
-    setForm({
-      type: "expense",
-      date: new Date().toISOString().slice(0, 10),
-      category: "",
-      amount: "",
-      note: "",
-    });
-    onCancel();
-  }
-
-  return (
-    <div className="max-w-2xl mx-auto">
-      <div className="bg-white rounded-xl shadow-lg border border-gray-100 overflow-hidden hover:shadow-xl transition-shadow duration-300">
-        <div className="px-6 py-4 border-b border-gray-100">
-          <h2 className="text-lg font-semibold text-gray-900">
-            {editingTransaction ? "Edit Transaction" : "Add New Transaction"}
-          </h2>
-          <p className="text-sm text-gray-500 mt-1">
-            {editingTransaction
-              ? "Update the transaction details"
-              : "Record a new income or expense"}
-          </p>
-        </div>
-        <div className="p-6 space-y-6">
-          <div className="grid grid-cols-2 gap-6">
-            <div className="space-y-2">
-              <label className="block text-sm font-semibold text-gray-700">
-                Transaction Type
-              </label>
-              <div className="relative">
-                <select
-                  className="w-full py-4 px-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-lg bg-white transition-all duration-200 appearance-none font-medium"
-                  value={form.type}
-                  onChange={(e) =>
-                    setForm((f) => ({ ...f, type: e.target.value }))
-                  }
-                >
-                  <option value="expense">💸 Expense</option>
-                  <option value="income">💰 Income</option>
-                </select>
-                <div className="absolute inset-y-0 right-0 flex items-center pr-3 pointer-events-none">
-                  <svg
-                    className="h-5 w-5 text-gray-400"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M19 9l-7 7-7-7"
-                    />
-                  </svg>
-                </div>
-              </div>
-            </div>
-            <div className="space-y-2">
-              <label className="block text-sm font-semibold text-gray-700">
-                Date
-              </label>
-              <input
-                type="date"
-                className="w-full py-4 px-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-lg bg-white transition-all duration-200"
-                value={form.date}
-                onChange={(e) =>
-                  setForm((f) => ({ ...f, date: e.target.value }))
-                }
-              />
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <label className="block text-sm font-semibold text-gray-700">
-              Category
-            </label>
-            <input
-              type="text"
-              placeholder="e.g., Feed, Fertilizer, Milk Sales, Crop Sales"
-              className="w-full py-4 px-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-lg bg-white transition-all duration-200"
-              value={form.category}
-              onChange={(e) =>
-                setForm((f) => ({ ...f, category: e.target.value }))
-              }
-            />
-          </div>
-
-          <div className="space-y-2">
-            <label className="block text-sm font-semibold text-gray-700">
-              Amount (LKR)
-            </label>
-            <div className="relative">
-              <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                <span className="text-gray-500 text-lg">₨</span>
-              </div>
-              <input
-                type="number"
-                placeholder="0"
-                className="w-full pl-10 pr-4 py-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-lg bg-white transition-all duration-200"
-                value={form.amount}
-                onChange={(e) =>
-                  setForm((f) => ({ ...f, amount: e.target.value }))
-                }
-              />
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <label className="block text-sm font-semibold text-gray-700">
-              Notes (Optional)
-            </label>
-            <textarea
-              placeholder="Additional details about this transaction..."
-              className="w-full py-4 px-4 rounded-xl border-2 border-gray-200 focus:ring-2 focus:ring-emerald-500 focus:border-emerald-500 shadow-lg bg-white transition-all duration-200 resize-none"
-              rows={4}
-              value={form.note}
-              onChange={(e) => setForm((f) => ({ ...f, note: e.target.value }))}
-            />
-          </div>
-
-          <div className="flex items-center justify-between pt-6 border-t-2 border-gray-100">
-            <div className="flex space-x-4">
-              <button
-                type="button"
-                onClick={handleSubmit}
-                className="inline-flex items-center px-8 py-4 border border-transparent text-base font-bold rounded-xl text-white bg-gradient-to-r from-emerald-600 to-emerald-700 hover:from-emerald-700 hover:to-emerald-800 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-emerald-500 transition-all duration-200 shadow-xl hover:shadow-2xl transform hover:-translate-y-1"
-              >
-                {editingTransaction ? (
-                  <>
-                    <svg
-                      className="w-5 h-5 mr-2"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M4 16v1a3 3 0 003 3h10a3 3 0 003-3v-1m-4-8l-4-4m0 0L8 8m4-4v12"
-                      />
-                    </svg>
-                    Update Transaction
-                  </>
-                ) : (
-                  <>
-                    <svg
-                      className="w-5 h-5 mr-2"
-                      fill="none"
-                      viewBox="0 0 24 24"
-                      stroke="currentColor"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        strokeWidth={2}
-                        d="M12 4v16m8-8H4"
-                      />
-                    </svg>
-                    Add Transaction
-                  </>
-                )}
-              </button>
-              {editingTransaction && (
-                <button
-                  type="button"
-                  onClick={handleCancel}
-                  className="inline-flex items-center px-8 py-4 border-2 border-gray-300 text-base font-semibold rounded-xl text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-emerald-500 transition-all duration-200 shadow-lg hover:shadow-xl transform hover:-translate-y-1"
-                >
-                  <svg
-                    className="w-5 h-5 mr-2"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M6 18L18 6M6 6l12 12"
-                    />
-                  </svg>
-                  Cancel
-                </button>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default function FinanceDashboard() {
-  const [transactions, setTransactions] = useState([]);
-  const [activeTab, setActiveTab] = useState("overview");
-  const [editingTransaction, setEditingTransaction] = useState(null);
-
-  const tabs = [
-    {
-      id: "overview",
-      name: "Overview",
-      icon: (
-        <svg
-          className="w-5 h-5"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M9 19v-6a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2a2 2 0 002-2zm0 0V9a2 2 0 012-2h2a2 2 0 012 2v10m-6 0a2 2 0 002 2h2a2 2 0 002-2m0 0V5a2 2 0 012-2h2a2 2 0 012 2v14a2 2 0 01-2 2h-2a2 2 0 01-2-2z"
-          />
-        </svg>
-      ),
-    },
-    {
-      id: "transactions",
-      name: "Transactions",
-      icon: (
-        <svg
-          className="w-5 h-5"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M9 5H7a2 2 0 00-2 2v10a2 2 0 002 2h8a2 2 0 002-2V7a2 2 0 00-2-2h-2M9 5a2 2 0 002 2h2a2 2 0 002-2M9 5a2 2 0 012-2h2a2 2 0 012 2"
-          />
-        </svg>
-      ),
-    },
-    {
-      id: "add",
-      name: editingTransaction ? "Edit" : "Add New",
-      icon: (
-        <svg
-          className="w-5 h-5"
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={2}
-            d="M12 4v16m8-8H4"
-          />
-        </svg>
-      ),
-    },
-  ];
-
-  function handleSave(transaction, action) {
-    if (action === "edit") {
-      setTransactions((prev) =>
-        prev.map((t) => (t.id === transaction.id ? transaction : t))
-      );
-    } else {
-      setTransactions((prev) => [transaction, ...prev]);
-    }
-    setEditingTransaction(null);
-    setActiveTab("transactions");
-  }
-
-  function handleEdit(transaction) {
-    setEditingTransaction(transaction);
-    setActiveTab("add");
-  }
-
-  function handleDelete(id) {
-    setTransactions((prev) => prev.filter((t) => t.id !== id));
-  }
-
-  function handleAddNew() {
-    setEditingTransaction(null);
-    setActiveTab("add");
-  }
-
-  function handleCancel() {
-    setEditingTransaction(null);
-    setActiveTab("transactions");
-  }
-
-  return (
-    <div className="min-h-screen bg-gradient-to-br from-gray-50 via-white to-emerald-50">
-      {/* Header */}
-      <div className="bg-white border-b border-gray-200 sticky top-0 z-40 backdrop-blur-sm bg-white/95 shadow-lg">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex items-center justify-between h-16">
-            <div className="flex items-center space-x-4">
-              <div className="h-8 w-8 rounded-lg bg-gradient-to-br from-emerald-500 to-emerald-600 flex items-center justify-center">
-                <svg
-                  className="w-5 h-5 text-white"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M12 8c-1.657 0-3 .895-3 2s1.343 2 3 2 3 .895 3 2-1.343 2-3 2m0-8c1.11 0 2.08.402 2.599 1M12 8V7m0 1v8m0 0v1m0-1c-1.11 0-2.08-.402-2.599-1"
-                  />
-                </svg>
-              </div>
-              <div>
-                <h1 className="text-xl font-semibold text-gray-900">
-                  Finance Dashboard
-                </h1>
-                <p className="text-sm text-gray-500">
-                  Track your farm's financial performance
-                </p>
-              </div>
-            </div>
-            <div className="flex items-center space-x-4">
-              <a
-                href="/admin/payroll"
-                className="inline-flex items-center px-4 py-2 bg-emerald-600 hover:bg-emerald-700 text-white rounded-lg text-sm font-medium transition-colors"
-              >
-                <svg
-                  className="w-4 h-4 mr-2"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M17 9V7a2 2 0 00-2-2H5a2 2 0 00-2 2v6a2 2 0 002 2h2m2 4h10a2 2 0 002-2v-6a2 2 0 00-2-2H9a2 2 0 00-2 2v6a2 2 0 002 2zm7-5a2 2 0 11-4 0 2 2 0 014 0z"
-                  />
-                </svg>
-                Payroll
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      {/* Tab Navigation */}
-      <div className="bg-white border-b border-gray-200 shadow-md">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex space-x-8">
-            {tabs.map((tab) => (
-              <button
-                key={tab.id}
-                onClick={() => setActiveTab(tab.id)}
-                className={`flex items-center space-x-2 py-4 px-1 border-b-2 font-medium text-sm transition-colors ${
-                  activeTab === tab.id
-                    ? "border-emerald-500 text-emerald-600"
-                    : "border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300"
-                }`}
-              >
-                {tab.icon}
-                <span>{tab.name}</span>
-              </button>
-            ))}
-          </div>
-        </div>
-      </div>
-
-      {/* Content */}
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
-        {activeTab === "overview" && (
-          <OverviewPage transactions={transactions} />
-        )}
-
-        {activeTab === "transactions" && (
-          <TransactionsPage
-            transactions={transactions}
-            onEdit={handleEdit}
-            onDelete={handleDelete}
-            onAddNew={handleAddNew}
-          />
-        )}
-
-        {activeTab === "add" && (
-          <AddNewPage
-            editingTransaction={editingTransaction}
-            onSave={handleSave}
-            onCancel={handleCancel}
-          />
-        )}
-      </div>
-    </div>
-  );
-}
